conversions stNEAR<->usd

+++ usdnear/src/utils.rs
@@ -39,7 +39,7 @@ pub fn is_promise_success() -> bool {
     }
 }
 
-pub fn apply_pct(basis_points:u16, amount:u128) -> u128 {
+pub fn apply_pct(basis_points:u32, amount:u128) -> u128 {
     return (U256::from(basis_points) * U256::from(amount) / U256::from(10_000)).as_u128() ;
 }
 pub fn apply_multiplier(amount:u128, basis_points:u16) -> u128 {
